refactor(avatars): address review — media-scoped reports, flags trait
- ReportUserAvatar: scope the pending-report idempotency check to the
  reported media_id so a swapped avatar is reportable on its own rather
  than folded into the stale pending report (+ regression test).
- Extract HasModerationFlags trait from User for the generic
  has/set/clear flag methods; User uses it. getAvatarUploadBlocked stays
  as the @method GraphQL seam, now delegating to the trait.
